Hash table variants and optimizations

Introduces an alternative F14 style interleave layour for
HashTable. This layout alternates tags and pointers to payload. Each
group takes two cache lines: 16 bytes of tags, 8 * 48 bytes of
pointers in the first cache line and 8*48 bits of pointers in the
second cache line. A hash table that is expected to hit frequently,
i.e. a group by table, will benefit from this layout since the tags
and pointers most often share one cache miss and always share the TLB
lookup.

Adds optional counters for hash lookups and collisions.

simplifies the hashProbe concept by passing the table as an argument
as opposed to different fields of the table.


Uses SIMD for an array hash table probe. Previously this was limited
to group by tables.



Before this change, optimized build on devserver

TPCH 30G dataset in Parquet.
./tpch --data_path=tpchpq30links  --num_splits_per_file=4 --num_drivers=14

velox/benchmarks/tpch/TpchBenchmark.cpp     relative  time/iter   iters/s
============================================================================
q1                                                           1.27s   785.53m
q3                                                           1.95s   512.91m
q5                                                           2.17s   461.62m
q6                                                           1.39s   720.91m
q7                                                           2.30s   435.01m
q8                                                           2.35s   424.85m
q9                                                           6.93s   144.33m
q10                                                          5.96s   167.90m
q12                                                          1.92s   520.29m
q13                                                         10.47s    95.55m
q14                                                          2.22s   450.48m
q15                                                          2.04s   489.87m
q16                                                       998.81ms      1.00
q18                                                          3.97s   251.97m
q19                                                          2.15s   465.10m
q22                                                          3.04s   329.25m

real	0m52.517s
user	5m4.079s
sys	1m41.267s

After this change, all hash tables interleaved

q1                                                           1.31s   761.65m
q3                                                           1.92s   521.60m
q5                                                           2.53s   395.92m
q6                                                        369.27ms      2.71
q7                                                           1.34s   745.68m
q8                                                           2.42s   413.91m
q9                                                           6.12s   163.30m
q10                                                          6.34s   157.69m
q12                                                          1.97s   508.39m
q13                                                          8.73s   114.58m
q14                                                          2.28s   439.46m
q15                                                          1.06s   939.13m
q16                                                          1.28s   782.58m
q18                                                          3.98s   251.51m
q19                                                          2.13s   469.96m
q22                                                          2.87s   348.84m

real	0m47.766s
user	4m55.719s
sys	1m49.795s