@oerling 2022-11-08
The loop conditionmeans that I + 7 is in range. It is not clear what you mean by batch sizes 2 or 4. Less than 8 is not considered here because it is considered in the next loop.  The test cases have every kind of non-multiple of 8 number of inputs.