@Yuhta 2022-11-08
@oerling On AVX2 we issue 2 gathers to cover the 8 words.  On SSE or NEON you would need 4 gathers.